# perf(html/ast): Use `Atom` instead of `JsWord`