--- a/crates/swc_common/Cargo.toml
+++ b/crates/swc_common/Cargo.toml
@@ -48,6 +48,7 @@ serde = { version = "1.0.119", features = ["derive"] }
 siphasher = "0.3.9"
 sourcemap = { version = "6", optional = true }
 string_cache = "0.8.4"
+swc_atoms = { version = "0.2.12", path = "../swc_atoms" }
 swc_eq_ignore_macros = { version = "0.1", path = "../swc_eq_ignore_macros" }
 swc_visit = { version = "0.3.0", path = "../swc_visit" }
 termcolor = { version = "1.0", optional = true }