**Description:**



**Related issue (if exists):**


---



# `main`


```
html/parser_document/css_2021_spec
                        time:   [14.183 ms 14.218 ms 14.261 ms]
Found 14 outliers among 100 measurements (14.00%)
  5 (5.00%) high mild
  9 (9.00%) high severe

html/parser_document/github_com_17_05_2022
                        time:   [5.4104 ms 5.4194 ms 5.4321 ms]
Found 13 outliers among 100 measurements (13.00%)
  6 (6.00%) high mild
  7 (7.00%) high severe

html/parser_document/stackoverflow_com_17_05_2022
                        time:   [4.9527 ms 4.9573 ms 4.9625 ms]
Found 13 outliers among 100 measurements (13.00%)
  1 (1.00%) low mild
  5 (5.00%) high mild
  7 (7.00%) high severe

html/parser_document_fragment/css_2021_spec
                        time:   [14.199 ms 14.219 ms 14.247 ms]
Found 15 outliers among 100 measurements (15.00%)
  5 (5.00%) high mild
  10 (10.00%) high severe

html/parser_document_fragment/github_com_17_05_2022
                        time:   [5.3940 ms 5.3984 ms 5.4035 ms]
Found 12 outliers among 100 measurements (12.00%)
  4 (4.00%) high mild
  8 (8.00%) high severe

html/parser_document_fragment/stackoverflow_com_17_05_2022
                        time:   [4.9630 ms 4.9712 ms 4.9813 ms]
Found 14 outliers among 100 measurements (14.00%)
  6 (6.00%) high mild
  8 (8.00%) high severe
```