@kdy1 2022-06-29
cc @alexander-akait 
The performance is regressed.

I think it will show better performance in real world, but there's one more thing we should try.
We should try expanding `static` set of `swc_atoms`